Allows effects that can be zoomed to be compiled.
The 'wrapSurf' (in module Shady.CompileSurface) family of functions used to produce GLSL vertex
shaders that transformed (u,v) coordinates into a coordinate of form (x,y,z,1).

The '1' is actually a specific instance of the 'w' coordinate which can be thought of as a "zoom
factor". An easy way to zoom an effect is to change this value. A simple approach is to pass this
value in as a uniform. Thus, now we have a notion of *custom uniform* values (which can come from UI
elements) and *required uniform* values (such as zoom).

The 'wrapSurf' functions were modified to produce functions which take the zoom value as an argument
too.

Function 'compileEffect' (in module Shady.CompileEffect) was modified to reflect this
change.

It is now up to the display environment to pass in the zoom uniform value to the GLSL shaders.