More ways to manipulate objects
Objects can now be panned, zoomed, and rotated.
Had to add a few new functions for this
- translate      (which can be used with onX, onY, onZ)
- rotateByMatrix (which simply does a rotation by a rotation matrix).

The latter function I found a bit icky. Rotation matrices don't feel
nearly elegant enough for the Shady language. However, in an
interactive environment it makes sense (for ease of use) to rotate with
respect to the current orientation of the object.

Shady objects do not have a notion of "current orientation". Thus we store
the current orientation in the display environment (which for me is
implemented in Javascript). A rotation matrix is a convenient representation
for this.